feat(neovim):Update nvim LSP configurations in dein_lazy.toml (#143)
## [optional body]
This commit updates the LSP configurations in the dein_lazy.toml file for Neovim. Specifically, it changes the autocmd BufWritePre to call vim.lsp.buf.format instead of vim.lsp.buf.formatting_sync, which should be faster. It also adds key mappings for code actions and references. No changes were made to the rest of the file.

## [optional footer(s)]