Feat/configure copilot model (#179)
## [optional body]
This pull request includes several changes to the configuration files
for Neovim and a new personality profile for an AI assistant. The most
important changes are the addition of a system prompt, modifications to
the extended prompts, and the introduction of a new personality profile.

### Configuration changes:

*
[`.config/nvim/copilot.toml`](diffhunk://#diff-f04922053842de2458379f6df241f2e77f615af8acfa165bcf45adc126860411R25-R35):
Added a new system prompt configuration to define the behavior and
personality of the CopilotChat assistant.
*
[`.config/nvim/copilot.toml`](diffhunk://#diff-f04922053842de2458379f6df241f2e77f615af8acfa165bcf45adc126860411L44-R55):
Updated the extended prompts to include additional instructions for test
planning and refactoring.
*
[`.config/nvim/copilot.toml`](diffhunk://#diff-f04922053842de2458379f6df241f2e77f615af8acfa165bcf45adc126860411R157-R165):
Added configurations for different models and updated the
`system_prompt` and `model` settings.

### Environment configuration:

*
[`.mise.toml`](diffhunk://#diff-1bac9641068b68e6b60d0166663560d45b271c427fd863f8d56ed36964f99ee2R3-R5):
Added a new environment configuration section to specify the `.env`
file.


## [optional footer(s)]